Makes make.sh fail if arg is not "uninstall"

+++ make.sh
@@ -2,11 +2,18 @@
 
 CURRDIR=$PWD
 
-if [ "$1" == "uninstall" ]
+if [ $# -gt 1 ]
 then
-    echo "Removing collection..."
-    rm -rfv ~/.ansible/collections/ansible_collections/andrewdomain
-    exit 0
+    if [ "$1" == "uninstall" ]
+    then
+	echo "Removing collection..."
+	rm -rfv ~/.ansible/collections/ansible_collections/andrewdomain
+	exit 0
+    else
+	echo 'you mean "uninstall" ?'
+	echo "got $@"
+	exit
+    fi
 fi
 
 cd `dirname -- ${BASH_SOURCE[0]}`
